requirements/base.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

# Bleeding edge Django
django==1.9.0

# Configuration
django-environ==0.4.0
django-secure==1.0.1
whitenoise==2.0.6
django-extensions==1.5.9

# Forms
django-braces==1.8.1
django-crispy-forms==1.5.2
django-floppyforms==1.6.0

# Models
django-model-utils==2.3.1

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.24.1

# Images
Pillow==2.9.0

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2015.7

# Redis support
django-redis==4.3.0
redis>=2.10.0